% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-GeM.R
\name{dVdt.GeM}
\alias{dVdt.GeM}
\title{Compute the VC Matrix}
\usage{
\method{dVdt}{GeM}(t, y, xds_obj, i)
}
\value{
a numeric \link{matrix}
}
\description{
Compute vectorial capacity for
the \code{GeM} module
}
\keyword{internal}
