% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-macdonald.R
\name{dVdt.macdonald}
\alias{dVdt.macdonald}
\title{Compute the VC Matrix}
\usage{
\method{dVdt}{macdonald}(t, y, xds_obj, i)
}
\value{
a numeric \link{matrix}
}
\description{
Compute vectorial capacity for
the \code{macdonald} module
}
\keyword{internal}
