% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-SI.R
\name{dVdt.SI}
\alias{dVdt.SI}
\title{Compute the VC Matrix}
\usage{
\method{dVdt}{SI}(t, y, xds_obj, i)
}
\value{
a numeric \link{matrix}
}
\description{
Compute vectorial capacity for
the \code{SI} module
}
\keyword{internal}
