% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-plots.R
\name{plot_connectivity_sankey}
\alias{plot_connectivity_sankey}
\title{Plot Connectivity as a Sankey Diagram}
\usage{
plot_connectivity_sankey(M, clrs, oclrs = NULL)
}
\arguments{
\item{M}{a connectivity matrix}

\item{clrs}{colors for the patches}

\item{oclrs}{opaque colors for the flows}
}
\value{
invisible
}
\description{
Plot Connectivity as a Sankey Diagram
}
