% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_xy.R
\name{plot_flow_sankey}
\alias{plot_flow_sankey}
\title{Plot Connectivity as a Sankey Diagram}
\usage{
plot_flow_sankey(M, clrs, oclrs = NULL)
}
\arguments{
\item{M}{an asymmetric matrix}

\item{clrs}{colors for the patches}

\item{oclrs}{opaque colors for the flows}
}
\value{
invisible
}
\description{
Plot Connectivity as a Sankey Diagram
}
