% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity-plots.R
\name{plot_kernel_graph}
\alias{plot_kernel_graph}
\title{Plot Kernel}
\usage{
plot_kernel_graph(
  M,
  xy = NULL,
  clrs = "black",
  scl = 1,
  labels = NULL,
  lab.cex = 0.6,
  arr.pos = 0.33,
  mn = 0.02
)
}
\arguments{
\item{M}{a connectivity matrix}

\item{xy}{locations of the patches}

\item{clrs}{colors}

\item{scl}{scaling for graph thickness}

\item{labels}{a set of labels}

\item{lab.cex}{size of the labels}

\item{arr.pos}{position of the arrow}

\item{mn}{plot if weight is greater than mn}
}
\value{
invisible
}
\description{
Plot Kernel
}
