% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_analyze.R
\name{xds_analyze}
\alias{xds_analyze}
\title{Analyze a Dynamical System}
\usage{
xds_analyze(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Compute the vectorial capacity and human transmitting capacity matrices.
}
