% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xds_scaling.eir}
\alias{xds_scaling.eir}
\title{Compute eir-pr scaling relationships}
\usage{
\method{xds_scaling}{eir}(xds_obj, N = 25)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}
}
\value{
an \strong{\code{xds}} model object
}
\description{
This function calls \link[ramp.xds:xds_solve]{ramp.xds::xds_solve} computes average annual values for the eir, the pr, and other
interesting terms and returns a table. It is computed for a xds_obj of class "cohort"
}
\keyword{internal}
