% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{xds_scaling.Lambda}
\alias{xds_scaling.Lambda}
\title{Compute scaling relationships from mosquito emergence through PfPR}
\usage{
\method{xds_scaling}{Lambda}(xds_obj, N = 30)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{N}{an integer}
}
\value{
\strong{\code{xds}} xds_obj object
}
\description{
This function calls \link[ramp.xds:xds_solve]{ramp.xds::xds_solve} to get the scaling
relationships for mosquito density over 9 factors of 10, and average annual
values for the eir, the pr, and other
interesting terms and returns a table.
}
\keyword{internal}
